shipit/api: renaming from shipit-workflow
