+++ lib/please_cli/please_cli/config.py
@@ -717,7 +717,7 @@ PROJECTS_CONFIG = {
             },
         ],
     },
-    'shipit-workflow': {
+    'shipit/api': {
         'checks': [
             ('Checking code quality', 'flake8'),
             ('Running tests', 'pytest tests/'),
@@ -731,36 +731,18 @@ PROJECTS_CONFIG = {
         ],
         'deploys': [
             {
-                'target': 'HEROKU',
-                'options': {
-                    'testing': {
-                        'nix_path_attribute': 'docker',
-                        'heroku_app': 'shipit-testing-workflow',
-                        'heroku_dyno_type': 'web',
-                        'url': 'https://shipit-workflow.testing.mozilla-releng.net',
-                        # TODO: we need to change this to SSL Endpoint
-                        'dns': 'shipit-workflow.testing.mozilla-releng.net.herokudns.com',
-                    },
-                    'staging': {
-                        'nix_path_attribute': 'docker',
-                        'heroku_app': 'shipit-staging-workflow',
-                        'heroku_dyno_type': 'web',
-                        'url': 'https://shipit-workflow.staging.mozilla-releng.net',
-                        # TODO: we need to change this to SSL Endpoint
-                        'dns': 'shipit-workflow.staging.mozilla-releng.net.herokudns.com',
-                    },
-                },
-            },
-            {
                 'target': 'DOCKERHUB',
                 'options': {
                     'testing': {
+                        'url': 'https://api.shipit.testing.mozilla-releng.net',
                         'nix_path_attribute': 'dockerflow',
                     },
                     'staging': {
+                        'url': 'https://api.shipit.staging.mozilla-releng.net',
                         'nix_path_attribute': 'dockerflow',
                     },
                     'production': {
+                        'url': 'https://api.shipit.mozilla-releng.net',
                         'nix_path_attribute': 'dockerflow',
                     },
                 },
@@ -776,7 +758,7 @@ PROJECTS_CONFIG = {
             },
         },
         'requires': [
-            'shipit-workflow',
+            'shipit/api',
         ],
         'checks': [
             ('Checking code quality', 'yarn lint'),
