+++ nix/setup.cfg
@@ -36,7 +36,7 @@ known_first_party =
     shipit_static_analysis
     shipit_taskcluster
     shipit_uplift
-    shipit_workflow
+    shipit_api
 
 [mypy]
 # Specify the target platform details in config, so your developers are
