+++ src/shipit/api/requirements_override.nix
@@ -1,0 +1,1 @@
+../../../nix/requirements_override.nix
