+++ src/shipit/api/setup.cfg
@@ -1,0 +1,1 @@
+../../../nix/setup.cfg
