+++ src/shipit/api/shipit_api/config.py
@@ -3,6 +3,5 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import shipit_workflow
-
-app = shipit_workflow.create_app()
+PROJECT_NAME = 'shipit-api'
+APP_NAME = 'shipit_api'
