+++ src/shipit/api/shipit_api/partners.py
