+++ src/shipit/frontend/README.md
@@ -3,7 +3,7 @@ Ship-It v2 UI
 
 To run a local version at https://localhost:8010/:
     $ yarn
-    # To run against staging API at https://shipit-workflow.staging.mozilla-releng.net/
+    # To run against staging API at https://api.shipit.staging.mozilla-releng.net/
     $ yarn start
     # To run against local API at https://localhost:8015
     $ CONFIG=dev yarn start
