--- src/shipit_workflow/requirements.txt
@@ -1,8 +1,0 @@
--e ./../../lib/cli_common[log] #egg=mozilla-cli-common
--e ./../../lib/backend_common[log,security,cors,api,auth0,db] #egg=mozilla-backend-common
-
-json-e
-
-# deployment
-gunicorn
-psycopg2
