--- src/shipit_workflow/setup.cfg
@@ -1,1 +1,0 @@
-../../nix/setup.cfg
